% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat.R
\name{compustat}
\alias{compustat}
\title{Download Compustat}
\usage{
compustat(wrds, begin_year, end_year, frequency, vars = "default")
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{begin_year}{Numeric.}

\item{end_year}{Numeric.}

\item{frequency}{Character. Input either 'annual' or 'quarterly'.}

\item{vars}{Optional character. Defaults to a recommended subset of variables. Choose 'all' if
you want to download all variables. Otherwise, provide a list of variables.}
}
\description{
\code{compustat} downloads the annual or quarterly Compustat table.
}
\details{
Downloads the annual or quarterly Compustat table. The table address is either compa.funda or
compa.fundq. By default, the function downloads a subset of the variables. You can specify your
own variable list by providing a string vector to the \code{vars} function. You can download all
the variables by setting \code{vars} equal to 'all'.  By default, mergeos on industry
information from the names table and removes duplicate gvkey fiscal quarter or
fiscal year observations. The de-duping logic is to keep the observation with the fewest NA
values. See function documentation for \code{compustat_annual} and \code{compustat_quarterly}.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
compa_df = compustat(wrds = wrds, begin_year = 2010, end_year = 2012, frequency = 'annual', vars = c('gvkey','fyear', 'revt'))
}
