% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat.R
\name{compustat_annual}
\alias{compustat_annual}
\title{Download annual Compustat}
\usage{
compustat_annual(wrds, begin_year, end_year, subset = TRUE)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{begin_year}{Numeric.}

\item{end_year}{Numeric.}

\item{subset}{Optional Boolean. Download recommended subset of variables? Defaults to \code{TRUE}.}
}
\description{
\code{compustat_annual} downloads the Compustat table with annual observations.
}
\details{
Downloads the annual observation Compustat table that is updated at an annual frequency. The
table address is compa.funda. By default, the function only downloads a subset of the variables.
You can download all the variables by setting the \code{subset} argument to \code{FALSE}.
By default, merges on industry information from the names table and removes duplicate gvkey
fiscal year observations. The de-duping logic is to keep the observation with the fewest NA
values.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
compa_df = compustat_annual(wrds = wrds, begin_year = 2010, end_year = 2012)
}
