% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat_crsp.R
\name{compustat_append_crsp_links}
\alias{compustat_append_crsp_links}
\title{Append CRSP identifiers onto a Compustat data frame}
\usage{
compustat_append_crsp_links(wrds, comp_df)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{comp_df}{Compustat data.frame. Requires the gkvey and datadate fields to be present.}
}
\description{
\code{compustat_append_crsp_links} appends on CRSP permno and permco identifiers onto Compustat
gvkey identifiers.
}
\details{
Appends on CRSP permno and permcos based on the standard Compsutat-CRSP linking table. The
linking table address is crspa.ccmxpf_linktable. Requires that the Compustat data.frame be
physically located in memory (not a lazy \code{dplyr} table reference). Requires that the gvkey
and datadate fields are present in the Compustat data.frame.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
compa_annual = compustat_annual(wrds = wrds, begin_year = 2010, end_year = 2011)
comp_linked = compustat_append_crsp_links(wrds = wrds, comp_df = compa_annual)
}
