% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat_crsp.R
\name{compustat_crsp_annual}
\alias{compustat_crsp_annual}
\title{Download annual Compustat with CRSP annual returns}
\usage{
compustat_crsp_annual(wrds, begin_year, end_year, subset = TRUE)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{begin_year}{Numeric.}

\item{end_year}{Numeric.}

\item{subset}{Optional Boolean. Download recommended subset of variables? Defaults to \code{TRUE}.}
}
\description{
\code{compustat_annual} downloads the annual Compustat table and merges on annual CRSP returns.
}
\details{
Downloads the annual Compustat file using the \code{compustat_annual} function. Appends on
CRSP links and returns using the functions \code{compustat_append_crsp_links} and
\code{crsp_annual}, respectively. The annual return merge is naive and is based on a fiscal year
to calendar year match.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
comp_crsp_df = compustat_crsp_annual(wrds = wrds, begin_year = 2010, end_year = 2012)
}
