% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat_crsp.R
\name{compustat_crsp_linking_table}
\alias{compustat_crsp_linking_table}
\title{Download Compustat-CRSP linking table}
\usage{
compustat_crsp_linking_table(wrds, dl = TRUE)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{dl}{Optional Boolean. Download the data? Defaults to \code{TRUE}. \code{FALSE} outputs a
lazy \code{dplyr} table reference.}
}
\description{
\code{compustat_crsp_linking_table} downloads the Compustat-CRSP linking table.
}
\details{
Downloads the standard Compustat-CRSP linking table. The table address is crspa.ccmxpf_linktable.
By default, the function downloads the data. A lazy \code{dplyr} table reference can be returned
by setting the \code{dl} argument to \code{FALSE}.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
linking_tbl = compustat_crsp_linking_table(wrds = wrds, dl = FALSE)
}
