% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compustat.R
\name{compustat_names}
\alias{compustat_names}
\title{Download Compustat names table}
\usage{
compustat_names(wrds, subset = TRUE, dl = TRUE)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{subset}{Optional Boolean. Download recommended subset of variables? Defaults to \code{TRUE}.}

\item{dl}{Optional Boolean. Download the data? Defaults to \code{TRUE}. \code{FALSE} outputs a
lazy \code{dplyr} table reference.}
}
\description{
\code{compustat_names} downloads the Compustat names table.
}
\details{
The names table contains general gvkey-level information, such as NAICS industry. The table
address is compa.names. By default, the function only downloads a subset of the variables.
You can download all the variables by setting the \code{subset} argument to \code{FALSE}.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
names_df = compustat_names(wrds = wrds, subset = TRUE, dl = TRUE)
}
