% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsp.R
\name{crsp}
\alias{crsp}
\title{Download CRSP}
\usage{
crsp(wrds, begin_year, end_year, frequency, dl = FALSE)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{begin_year}{Numeric.}

\item{end_year}{Numeric.}

\item{frequency}{Character. Input either 'annual', 'monthly', or 'daily'.}

\item{dl}{Optional Boolean. Download the data? Defaults to \code{TRUE}. \code{FALSE} outputs a
lazy \code{dplyr} table reference.}
}
\description{
\code{crsp} downloads the annual, monthly, or daily CRSP table.
}
\details{
Downloads the annual, monthly, or daily CRSP Compustat table. The schema is crspa and the
table is either the aggregated msf, msf, or dsf. Removes observations with weakly negative prices
and shares outstanding. The function merges on identifying information from the
CRSP header table. Please see either \code{crsp_annual}, \code{crsp_monthly}, or
\code{crsp_daily} documentation for more details.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
crsp_df = crsp(wrds = wrds, begin_year = 2010, end_year = 2012, frequency = 'annual')
}
