% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsp.R
\name{crsp_header}
\alias{crsp_header}
\title{Download CRSP header table}
\usage{
crsp_header(wrds, subset = TRUE, dl = TRUE)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{subset}{Optional Boolean. Download recommended subset of variables? Defaults to \code{TRUE}.}

\item{dl}{Optional Boolean. Download the data? Defaults to \code{TRUE}. \code{FALSE} outputs a
lazy \code{dplyr} table reference.}
}
\description{
\code{crsp_header} downloads the CRSP header table.
}
\details{
The header table contains permno-level information, such as SIC industry. The table
address is crspa.msfhdr. By default, the function only downloads a subset of the variables.
You can download all the variables by setting the \code{subset} argument to \code{FALSE}.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
crsp_names = crsp_header(wrds = wrds, subset = TRUE, dl = FALSE)
}
