% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsp.R
\name{crsp_monthly}
\alias{crsp_monthly}
\title{Download CRSP monthly return and price data}
\usage{
crsp_monthly(wrds, begin_year, end_year, dl = TRUE)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{begin_year}{Numeric.}

\item{end_year}{Numeric.}

\item{dl}{Optional Boolean. Download the data? Defaults to \code{TRUE}. \code{FALSE} outputs a
lazy \code{dplyr} table reference.}

\item{subset}{Optional Boolean. Download recommended subset of variables? Defaults to \code{TRUE}.}
}
\description{
\code{crsp_monthly} downloads CRSP monthly returns and end-of-month prices.
}
\details{
Downloads data from table crspa.msf. Removes observations with weakly negative prices and
shares outstanding. The function merges on identifying information from the
CRSP header table.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
crspm_df = crsp_monthly(wrds = wrds, begin_year = 2010, end_year = 2012)
}
