% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergent.R
\name{mergent_corporates}
\alias{mergent_corporates}
\title{Download Mergent corporate bonds}
\usage{
mergent_corporates(wrds, clean = TRUE, vanilla = TRUE, subset = TRUE,
  dl = TRUE)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{clean}{Optional Boolean. Clean the data to remove data errors and missings? Defaults to
\code{TRUE}.}

\item{vanilla}{Optional Boolean. Restrict data to common vanilla bonds? Defaults to \code{TRUE}.}

\item{subset}{Optional Boolean. Clean the data for errors and missing fields? Defaults to
\code{TRUE}.}

\item{dl}{Optional Boolean. Download the data? Defaults to \code{TRUE}. \code{FALSE} outputs a
lazy \code{dplyr} table reference.}
}
\description{
\code{mergent_corporates} downloads corporate bond issues from the Mergent issues table
}
\details{
Downloads issue-level bond data from the table fisd.fisd_mergedissue for corporate bonds.
Corporates are defined by being in utility, financial, or industrial Mergent industry groups,
having bond types in CDEB, CMTN, CZ, CMTZ, UCID, or CP, and having security levels in
SEN, SENS, SUB, or SS. This function calls the \code{mergent_issues} function to download the
data You can download all the variables by setting the \code{subset}
argument to \code{FALSE}. By default, this function also cleans the data and restricts to
vanilla corporates, which removes: floating rates notes, convertible bonds, foreign domiciled
issuers, foreign currency notes, private placements, perpetuals, preferred securities,
retail notes, slobs, exchangeables, unit deals, pay-in-kinds, defeased bonds, and bonds with
non-strictly positive offering amounts and missing coupon types, issue ids, maturity, and
offering date. By default, also merges on initial credit ratings for the issue and makes
simple variable transformations. Resticts to SEN, SENS, SUB, oR SS on campital structure,
although this is mostly without loss of generality (JUN, JUNS, NON comprise only 150 issues)
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
mergent_corps = mergent_corporates(wrds, clean = TRUE, vanilla = TRUE, subset = TRUE, dl = FALSE)
}
