% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergent.R
\name{mergent_issuer_panel}
\alias{mergent_issuer_panel}
\title{Mergent issuer panel summary}
\usage{
mergent_issuer_panel(wrds, begin_year, end_year, corps_only = FALSE,
  clean = FALSE, vanilla = FALSE, min_offering_amt = 1)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{begin_year}{Numeric.}

\item{end_year}{Numeric.}

\item{corps_only}{Boolean. Summarize corporate bonds only? Defaults to \code{FALSE}. If true,
function uses the filters in \code{mergent_corporates}.}

\item{clean}{Boolean. Restrict summary to bond issues with clean data? Clean is defined
as in the function \code{mergent_issues}.}

\item{vanilla}{Boolean. Restrict summary to vanilla bond issues only? Vanilla is defined
as in the function \code{mergent_issues}.}

\item{min_offering_amt}{Numeric. Mimum size (in thousands) for bonds to be included in the
summary.}
}
\description{
\code{mergent_issuer_panel} constructs a yearly panel of issuer-level bond summaries.
}
\details{
Aggregates issue-level bond data on an annual basis to the issuer level on an annual basis.
Summarizes several key features of each issuers bond issuance in each year, including
the number of bonds currently outstanding, the distribution of bond sizes, the average
bond ratings, the features of maturing bonds, and the features of newly issued bonds
within the year.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
mergent_df = mergent_issuer_panel(wrds = wrds, begin_year = 2000, end_year = 2002,
corps_only = TRUE, clean = TRUE, vanilla = TRUE, min_offering_amt = 1000)
}
