% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergent.R
\name{mergent_issues}
\alias{mergent_issues}
\title{Download Mergent bond issues table}
\usage{
mergent_issues(wrds, clean = TRUE, subset = TRUE, dl = TRUE)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{clean}{Optional Boolean. Clean the data to receive common bonds? Defaults to
\code{TRUE}.}

\item{subset}{Optional Boolean. Download recommended subset of variables? Defaults to
\code{TRUE}.}

\item{dl}{Optional Boolean. Download the data? Defaults to \code{TRUE}. \code{FALSE} outputs a
lazy \code{dplyr} table reference.}
}
\description{
\code{mergent_issues} downloads the Mergent bond issues table.
}
\details{
Downloads issue-level bond data from the table fisd.fisd_mergedissue. Appends on information
from the callable, agentid, and exchangeable tables. By default, the function only downloads a
subset of the variables. You can download all the variables by setting the \code{subset}
argument to \code{FALSE}. By default, this function also cleans the data of the following
types of bonds: foreign currency notes, private placements, perpetuals, preferred securities,
retail notes, slobs, exchangeables, unit deals, pay-in-kinds, defeased bonds, and bonds with
non-strictly positive offering amounts and missing coupon types, issue ids, maturity, and
offering date. By default, also merges on initial credit ratings for the issue and makes
simple variable transformations.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
mergent_issues_df = mergent_issues(wrds = wrds, clean = TRUE, subset = TRUE, dl = FALSE)
}
