% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonds_compustat_crsp.R
\name{mergent_linking_table}
\alias{mergent_linking_table}
\title{Download Mergent-CRSP-Compustat linking table}
\usage{
mergent_linking_table()
}
\description{
\code{mergent_linking_table} downloads the cleaned Mergent to CRSP-Compustat linking table.
}
\details{
Downloads the cleaned Mergent to CRSP-Compustat linking table from remote repository. Linking
table is created by cleaning the TRACE-CRSP linking table from WRDS and converting cusip to
permno links into Mergent issuer_id to permno and gvkey links. Performance of linking
table is suspect for issuers who have not issued a bond during existence of TRACE (2002+).
Linking table is constructed in create_dropboxdata.R.
}
\examples{
link_df = trace_crsp_linking_table()
}
