% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergent.R
\name{mergent_yearly_ao}
\alias{mergent_yearly_ao}
\title{Mergent yearly amount outstanding panel}
\usage{
mergent_yearly_ao(wrds, begin_year = 1995, end_year = 2019,
  dl = FALSE)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{begin_year}{Numeric. Default is 1995 (first year of updates to amount outstanding table).}

\item{end_year}{Numeric. Default is 2019.}

\item{dl}{Optional Boolean. Download the data? Defaults to \code{FALSE}.}
}
\description{
\code{mergent_yearly_ao} constructs a issue-year panel of amount outstandings.
}
\details{
Combines the Mergent issue table with the Mergent historical amount outstanding panel to
create a year-by-year summary of each issue's amount outstanding inclusive of bond event
(e.g., calling the bond). Includes yearly entries for issues only if their maturity is
less than the year of interest and their offering date is at least the year of interest.
Flags issue-year observations whenever there is zero estimated amount oustanding (e.g., the
bond has been called) but does not remove the observations. Construction is memory intensive
(several one-to-many merges) and requires support from the WRDS cloud. Constructs panel for
all bond issues and does not filter on any fields except positive initial amount oustanding
and non-missing maturity and offering years. Run-time is fast if WRDS cloud is leveraged.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
mergent_ao_panel = mergent_yearly_ao(wrds, begin_year = 1995, end_year = 2019, dl = TRUE)
table(mergent_ao_panel$year)
head(mergent_ao_panel \%>\% filter(issue_id == 1) \%>\% data.frame())
}
