% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergent.R
\name{mergent_yearly_ratings}
\alias{mergent_yearly_ratings}
\title{Mergent yearly credit ratings}
\usage{
mergent_yearly_ratings(wrds, begin_year = 1995, end_year = 2019,
  dl = FALSE)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{begin_year}{Numeric. Default is 1995 (first year of updates to ratings table).}

\item{end_year}{Numeric. Default is 2019.}

\item{dl}{Optional Boolean. Download the data? Defaults to \code{FALSE}.}
}
\description{
\code{mergent_yearly_ao} constructs an issue-year panel of S&P, Moodys, and Fitch ratings.
}
\details{
Constructs a year-by-year account of issue-level credit ratings from the Mergent issues table
and the historical credit ratings table. Considers only credit ratings from S&P, Moodys, and
Fitch. Only keeps issue observations between the offering and maturity date. Does not remove
a bond if the entire issue is called. Only filter applied is offering amounts above zero.
Assigns a numeric rating to each agency's letter rating. Lower numeric ratings signal higher
perceived credit quality; numeric ratings at-or-below 10 are considered investment grade.
Ratings are on a scale of 1 (highest perceived quality) to 22 (lowest perceived quality).
The following items are worth noting: (1) issues that are not rated by an agency are assigned
a rating of 'NR' for that agency and a numeric code of 23, (2) the Mergent database states that
the ratings data is only supported after 1995, despite there being pre-1995 ratings in the data.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
mergent_ratings_panel = mergent_yearly_ratings(wrds, begin_year = 1995, end_year = 2000,
dl = TRUE)
table(mergent_ratings_panel$year)
head(mergent_ratings_panel \%>\% filter(issue_id == 1) \%>\% data.frame())
}
