% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonds_compustat_crsp.R
\name{trace_crsp_linking_table}
\alias{trace_crsp_linking_table}
\title{Download TRACE-CRSP linking table}
\usage{
trace_crsp_linking_table()
}
\description{
\code{trace_crsp_linking_table} downloads the cleaned TRACE cusip to CRSP permno linking table.
}
\details{
Downloads the cleaned TRACE-CRSP linking table from remote repository. Linking table from
WRDS is cleaned using the procedure outlined in create_dropboxdata.R. Contact author for
access to this source code. Cleaning procedure removes erroneous links and extends link
start dates to prior to introduction of TRACE in 2002.
}
\examples{
link_df = trace_crsp_linking_table()
}
