% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_wrds.R
\name{wrds_connect}
\alias{wrds_connect}
\title{Connect to the WRDS cloud}
\usage{
wrds_connect(username, password)
}
\arguments{
\item{username}{Character. WRDS username.}

\item{password}{Character. WRDS password.}
}
\description{
\code{wrds_connect} creates a remote connection to the WRDS cloud.
}
\details{
Note: WRDS cloud runs PostgreSQL.
}
\examples{
wrds = wrds_connect(username = Sys.getenv("WRDS_NAME"), password = Sys.getenv("WRDS_PASS"))
}
