% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_wrds.R
\name{wrds_schema_list}
\alias{wrds_schema_list}
\title{List available schemas in WRDS.}
\usage{
wrds_schema_list(wrds)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}
}
\description{
\code{wrds_schema_list} downloads a list of the available schemas in the WRDS cloud.
}
\details{
Schemas represent different data vendors in WRDS. Schemas can be thought of as separate
databases that contain their own set of tables.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
schema_list = wrds_schema_list(wrds = wrds)
}
