% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_wrds.R
\name{wrds_table_list}
\alias{wrds_table_list}
\title{List the tables in a WRDS schema.}
\usage{
wrds_table_list(wrds, schema)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{schema}{Character. WRDS schema from \code{wrds_schema_list}.}
}
\description{
\code{wrds_table_list} lists the tables in a WRDS schema.
}
\details{
NA
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
table_list = wrds_table_list(wrds = wrds, schema = "compa")
}
