% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_wrds.R
\name{wrds_variable_list}
\alias{wrds_variable_list}
\title{List the variables in a WRDS table}
\usage{
wrds_variable_list(wrds, schema, table)
}
\arguments{
\item{wrds}{WRDS connection object from \code{wrds_connect} function.}

\item{schema}{Character. WRDS schema from \code{wrds_schema_list}.}

\item{table}{Character. WRDS table from \code{wrds_table_list}.}
}
\description{
\code{wrds_variable_list} lists the columns (variables) in a WRDS table.
}
\details{
NA.
}
\examples{
wrds = wrds_connect(username = "testing", password = "123456")
variable_list = wrds_variable_list(wrds = wrds, schema = "compa", table = "funda")
}
